perf(frontend): convert home page to server component for better SSR
